+++ verifyCredentials.js
@@ -11,7 +11,12 @@ module.exports = function verifyCredentials(credentials, cb) {
     }
     this.logger.trace('Current credentials: %j', credentials);
     const client = new Client(this.logger, credentials);
-    await client.listObjects({ Bucket: '' });
+    const bucketsNames = await client.listBucketNames();
+    if (bucketsNames.length < 1) {
+      this.logger.info('API keys are valid but they don\'t have permission to manipulate any existing buckets.');
+    } else {
+      this.logger.info(`The provided API keys have access to the following buckets: ${bucketsNames.join(', ')}`);
+    }
     this.logger.info('Verification succeeded');
     cb(null, { verified: true });
     return { verified: true };